Move Item Requirement code from EmoteClue to its own sub-package
