+++ runelite-client/src/main/java/net/runelite/client/plugins/cluescrolls/clues/item/AllRequirementsCollection.java
@@ -22,36 +22,26 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package net.runelite.client.plugins.cluescrolls.clues.emote;
+package net.runelite.client.plugins.cluescrolls.clues.item;
 
 import net.runelite.api.Client;
 import net.runelite.api.Item;
 
-public class RangeItemRequirement implements ItemRequirement
+public class AllRequirementsCollection implements ItemRequirement
 {
-	private String name;
-	private int startItemId;
-	private int endItemId;
+	private ItemRequirement[] requirements;
 
-	public RangeItemRequirement(String name, int startItemId, int endItemId)
+	public AllRequirementsCollection(ItemRequirement... requirements)
 	{
-		this.name = name;
-		this.startItemId = startItemId;
-		this.endItemId = endItemId;
+		this.requirements = requirements;
 	}
 
 	@Override
 	public boolean fulfilledBy(int itemId)
 	{
-		return itemId >= startItemId && itemId <= endItemId;
-	}
-
-	@Override
-	public boolean fulfilledBy(Item[] items)
-	{
-		for (Item item : items)
+		for (ItemRequirement requirement : requirements)
 		{
-			if (item.getId() >= startItemId && item.getId() <= endItemId)
+			if (requirement.fulfilledBy(itemId))
 			{
 				return true;
 			}
@@ -61,8 +51,22 @@ public class RangeItemRequirement implements ItemRequirement
 	}
 
 	@Override
+	public boolean fulfilledBy(Item[] items)
+	{
+		for (ItemRequirement requirement : requirements)
+		{
+			if (!requirement.fulfilledBy(items))
+			{
+				return false;
+			}
+		}
+
+		return true;
+	}
+
+	@Override
 	public String getCollectiveName(Client client)
 	{
-		return name;
+		return "N/A";
 	}
 }
