+++ runelite-client/src/main/java/net/runelite/client/plugins/cluescrolls/clues/item/ItemRequirement.java
@@ -22,53 +22,16 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package net.runelite.client.plugins.cluescrolls.clues.emote;
+package net.runelite.client.plugins.cluescrolls.clues.item;
 
 import net.runelite.api.Client;
 import net.runelite.api.Item;
 
-public class AnyRequirementCollection implements ItemRequirement
+public interface ItemRequirement
 {
-	private String name;
-	private ItemRequirement[] requirements;
+	boolean fulfilledBy(int itemId);
 
-	public AnyRequirementCollection(String name, ItemRequirement... requirements)
-	{
-		this.name = name;
-		this.requirements = requirements;
-	}
+	boolean fulfilledBy(Item[] items);
 
-	@Override
-	public boolean fulfilledBy(int itemId)
-	{
-		for (ItemRequirement requirement : requirements)
-		{
-			if (requirement.fulfilledBy(itemId))
-			{
-				return true;
-			}
-		}
-
-		return false;
-	}
-
-	@Override
-	public boolean fulfilledBy(Item[] items)
-	{
-		for (ItemRequirement requirement : requirements)
-		{
-			if (requirement.fulfilledBy(items))
-			{
-				return true;
-			}
-		}
-
-		return false;
-	}
-
-	@Override
-	public String getCollectiveName(Client client)
-	{
-		return name;
-	}
+	String getCollectiveName(Client client);
 }
