+++ runelite-client/src/main/java/net/runelite/client/plugins/cluescrolls/clues/item/RangeItemRequirement.java
@@ -22,51 +22,47 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package net.runelite.client.plugins.cluescrolls.clues.emote;
+package net.runelite.client.plugins.cluescrolls.clues.item;
 
 import net.runelite.api.Client;
-import net.runelite.api.EquipmentInventorySlot;
 import net.runelite.api.Item;
 
-public class SlotLimitationRequirement implements ItemRequirement
+public class RangeItemRequirement implements ItemRequirement
 {
-	private String description;
-	private EquipmentInventorySlot[] slots;
+	private String name;
+	private int startItemId;
+	private int endItemId;
 
-	public SlotLimitationRequirement(String description, EquipmentInventorySlot... slots)
+	public RangeItemRequirement(String name, int startItemId, int endItemId)
 	{
-		this.description = description;
-		this.slots = slots;
+		this.name = name;
+		this.startItemId = startItemId;
+		this.endItemId = endItemId;
 	}
 
 	@Override
 	public boolean fulfilledBy(int itemId)
 	{
-		return false;
+		return itemId >= startItemId && itemId <= endItemId;
 	}
 
 	@Override
 	public boolean fulfilledBy(Item[] items)
 	{
-		for (EquipmentInventorySlot slot : slots)
+		for (Item item : items)
 		{
-			if (slot.getSlotIdx() >= items.length)
+			if (item.getId() >= startItemId && item.getId() <= endItemId)
 			{
-				continue; //We can't check the slot, because there is nothing in it, the array hasn't been resized
-			}
-
-			if (items[slot.getSlotIdx()].getId() != -1)
-			{
-				return false;
+				return true;
 			}
 		}
 
-		return true;
+		return false;
 	}
 
 	@Override
 	public String getCollectiveName(Client client)
 	{
-		return description;
+		return name;
 	}
 }
