+++ runelite-client/src/main/java/net/runelite/client/plugins/cluescrolls/clues/item/SlotLimitationRequirement.java
@@ -22,40 +22,40 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package net.runelite.client.plugins.cluescrolls.clues.emote;
+package net.runelite.client.plugins.cluescrolls.clues.item;
 
 import net.runelite.api.Client;
+import net.runelite.api.EquipmentInventorySlot;
 import net.runelite.api.Item;
 
-public class AllRequirementsCollection implements ItemRequirement
+public class SlotLimitationRequirement implements ItemRequirement
 {
-	private ItemRequirement[] requirements;
+	private String description;
+	private EquipmentInventorySlot[] slots;
 
-	public AllRequirementsCollection(ItemRequirement... requirements)
+	public SlotLimitationRequirement(String description, EquipmentInventorySlot... slots)
 	{
-		this.requirements = requirements;
+		this.description = description;
+		this.slots = slots;
 	}
 
 	@Override
 	public boolean fulfilledBy(int itemId)
 	{
-		for (ItemRequirement requirement : requirements)
-		{
-			if (requirement.fulfilledBy(itemId))
-			{
-				return true;
-			}
-		}
-
 		return false;
 	}
 
 	@Override
 	public boolean fulfilledBy(Item[] items)
 	{
-		for (ItemRequirement requirement : requirements)
+		for (EquipmentInventorySlot slot : slots)
 		{
-			if (!requirement.fulfilledBy(items))
+			if (slot.getSlotIdx() >= items.length)
+			{
+				continue; //We can't check the slot, because there is nothing in it, the array hasn't been resized
+			}
+
+			if (items[slot.getSlotIdx()].getId() != -1)
 			{
 				return false;
 			}
@@ -67,6 +67,6 @@ public class AllRequirementsCollection implements ItemRequirement
 	@Override
 	public String getCollectiveName(Client client)
 	{
-		return "N/A";
+		return description;
 	}
 }
